+++ projects/geoprompt/src/geoprompt/__init__.py
@@ -1,13 +1,14 @@
-from .equations import area_similarity, coordinate_distance, corridor_strength, directional_alignment, directional_bearing, euclidean_distance, haversine_distance, prompt_decay, prompt_influence, prompt_interaction
+from .equations import accessibility_index, area_similarity, coordinate_distance, corridor_strength, directional_alignment, directional_bearing, euclidean_distance, gravity_model, haversine_distance, prompt_decay, prompt_influence, prompt_interaction
 from .frame import Bounds, GeoPromptFrame
-from .geometry import geometry_area, geometry_bounds, geometry_centroid, geometry_contains, geometry_distance, geometry_intersects, geometry_intersects_bounds, geometry_length, geometry_type, geometry_within, geometry_within_bounds, transform_geometry
+from .geometry import geometry_area, geometry_bounds, geometry_centroid, geometry_contains, geometry_convex_hull, geometry_distance, geometry_envelope, geometry_intersects, geometry_intersects_bounds, geometry_length, geometry_type, geometry_within, geometry_within_bounds, transform_geometry
 from .overlay import buffer_geometries, dissolve_geometries, geometry_from_shapely, geometry_to_geojson, geometry_to_shapely
-from .io import frame_to_geojson, read_features, read_geojson, read_points, write_geojson
+from .io import frame_to_geojson, frame_to_records_flat, read_features, read_geojson, read_points, write_geojson
 
 
 __all__ = [
     "Bounds",
     "GeoPromptFrame",
+    "accessibility_index",
     "area_similarity",
     "buffer_geometries",
     "coordinate_distance",
@@ -17,12 +18,15 @@ __all__ = [
     "directional_bearing",
     "euclidean_distance",
     "frame_to_geojson",
+    "frame_to_records_flat",
+    "geometry_convex_hull",
+    "geometry_distance",
+    "geometry_envelope",
     "geometry_from_shapely",
     "geometry_area",
     "geometry_bounds",
     "geometry_centroid",
     "geometry_contains",
-    "geometry_distance",
     "geometry_intersects",
     "geometry_intersects_bounds",
     "geometry_length",
@@ -31,6 +35,7 @@ __all__ = [
     "geometry_type",
     "geometry_within",
     "geometry_within_bounds",
+    "gravity_model",
     "haversine_distance",
     "read_geojson",
     "prompt_decay",